requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
python-dotenv